// threecommas-buythedip/types.ts
export interface DealType {
  id: number;
  type: string;
  bot_id: number;
  max_safety_orders: number;
  deal_has_error: boolean;
  from_currency_id: number;
  to_currency_id: number;
  account_id: number;
  active_safety_orders_count: number;
  created_at: string;
  updated_at: string;
  closed_at: null;
  "finished?": false;
  current_active_safety_orders_count: number;
  current_active_safety_orders: number;
  completed_safety_orders_count: number;
  completed_manual_safety_orders_count: number;
  "cancellable?": boolean;
  "panic_sellable?": boolean;
  trailing_enabled: boolean;
  tsl_enabled: boolean;
  stop_loss_timeout_enabled: boolean;
  stop_loss_timeout_in_seconds: number;
  active_manual_safety_orders: number;
  pair: string;
  status: string;
  localized_status: string;
  take_profit: string;
  base_order_volume: string;
  safety_order_volume: string;
  safety_order_step_percentage: string;
  leverage_type: string;
  leverage_custom_value: string;
  bought_amount: string;
  bought_volume: string;
  bought_average_price: string;
  base_order_average_price: string;
  sold_amount: string;
  sold_volume: string;
  sold_average_price: string;
  take_profit_type: string;
  final_profit: string;
  martingale_coefficient: string;
  martingale_volume_coefficient: string;
  martingale_step_coefficient: string;
  stop_loss_percentage: string;
  error_message: string;
  profit_currency: string;
  stop_loss_type: string;
  safety_order_volume_type: string;
  base_order_volume_type: string;
  from_currency: string;
  to_currency: string;
  current_price: string;
  take_profit_price: string;
  stop_loss_price: string;
  final_profit_percentage: string;
  actual_profit_percentage: string;
  bot_name: string;
  account_name: string;
  usd_final_profit: string;
  actual_profit: string;
  actual_usd_profit: string;
  failed_message: string;
  reserved_base_coin: string;
  reserved_second_coin: string;
  trailing_deviation: string;
  trailing_max_price: string;
  tsl_max_price: string;
  strategy: string;
}

export interface OrderType {
  order_id: string;
  order_type: string;
  deal_order_type: string;
  cancellable: boolean;
  status_string: string;
  created_at: string;
  updated_at: string;
  quantity: string;
  quantity_remaining: string;
  total: string;
  rate: string;
  average_price: string;
  error_description?: string;
  error?: string;
}

export interface SafetyOrderType {
  deal: DealType;
  order: OrderType;
}